backend/requirements/base.txt
Django
psycopg2
django-environ
djangorestframework
dj_rest_auth
django_extensions
ipython